Change type of local variables in siftDown function to remove warning about comparing int with size_t.

+++ PatientPriorityQueue.h
@@ -181,7 +181,7 @@ void PatientPriorityQueue::siftUp(int index) {
 }
 
 void PatientPriorityQueue::siftDown(int index) {
-    int leftIdx, rightIdx, minIdx;
+    size_t leftIdx, rightIdx, minIdx;
     leftIdx = getLeftChild(index);
     rightIdx = getRightChild(index);
